requirements.txt
discord.py
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2